+++ outputs.tf
@@ -111,6 +111,11 @@ output "cluster_nodes" {
   value       = try(aws_redshift_cluster.this[0].cluster_nodes, {})
 }
 
+output "cluster_namespace_arn" {
+  description = "The namespace Amazon Resource Name (ARN) of the cluster"
+  value       = try(aws_redshift_cluster.this[0].cluster_namespace_arn, null)
+}
+
 ################################################################################
 # Parameter Group
 ################################################################################
@@ -212,4 +217,22 @@ output "usage_limits" {
 output "authentication_profiles" {
   description = "Map of authentication profiles created and their associated attributes"
   value       = aws_redshift_authentication_profile.this
+}
+
+################################################################################
+# Managed Master Password Secret
+################################################################################
+
+output "master_password_secret_arn" {
+  description = "ARN of managed master password secret"
+  value       = try(aws_redshift_cluster.this[0].master_password_secret_arn, null)
+}
+
+################################################################################
+# Managed Secret Rotation
+################################################################################
+
+output "cluster_secretsmanager_secret_rotation_enabled" {
+  description = "Specifies whether automatic rotation is enabled for the secret"
+  value       = try(aws_secretsmanager_secret_rotation.this[0].rotation_enabled, null)
 }